Refactor: Improve stopwatch stop time accuracy
This commit refactors the `StopWatchViewModel` to calculate the stopwatch stop time more accurately.

Instead of using the current system time when stopping, the stop time is now calculated by adding the elapsed time from the `StopwatchState` to the `startMills` of the focus mode session.

A `toMillis()` function has been added to the `StopwatchState` data class to facilitate this calculation.